Update seo.php
Once the googletitle is set, there's no way for templates to access the original page title. There are many cases where the Google SEO title is too long and not suitable to display on the page as a title. This change creates a new displaytitle value to hold the original title before it's overwritten.